+++ CoachBot/Controllers/TournamentStaffController.cs
@@ -1,5 +1,6 @@
 ﻿using CoachBot.Domain.Model;
 using CoachBot.Domain.Services;
+using CoachBot.Extensions;
 using Microsoft.AspNetCore.Mvc;
 using static CoachBot.Attributes.HubRoleAuthorizeAttribute;
 
@@ -11,33 +12,54 @@ namespace CoachBot.Controllers
     public class TournamentStaffController : Controller
     {
         private readonly TournamentService _tournamentService;
+        private readonly PlayerService _playerService;
 
-        public TournamentStaffController(TournamentService tournamentService)
+        public TournamentStaffController(TournamentService tournamentService, PlayerService playerService)
         {
             _tournamentService = tournamentService;
+            _playerService = playerService;
         }
 
         [HubRolePermission(HubRole = PlayerHubRole.Administrator)]
         [HttpDelete("{id}")]
-        public void DeleteTournamentStaffMember(int id)
+        public IActionResult DeleteTournamentStaffMember(int id)
         {
+            var tournamentStaff = _tournamentService.GetTournamentStaffMember(id);
+            if (!_tournamentService.IsTournamentOrganiser((int)tournamentStaff.TournamentId, User.GetSteamId()) && !_playerService.IsOwner(User.GetSteamId()))
+            {
+                return Unauthorized();
+            }
+
             _tournamentService.DeleteTournamentStaff(id);
+
+            return Ok();
         }
 
         [HubRolePermission(HubRole = PlayerHubRole.Administrator)]
         [HttpPost]
-        public void CreateTournamentStaffMember(TournamentStaff tournamentStaff)
+        public IActionResult CreateTournamentStaffMember(TournamentStaff tournamentStaff)
         {
-            // TODO: MUST BE ADMIN OF TOURNAMENT
+            if (!_tournamentService.IsTournamentOrganiser((int)tournamentStaff.TournamentId, User.GetSteamId()) && !_playerService.IsOwner(User.GetSteamId()))
+            {
+                return Unauthorized();
+            }
+
             _tournamentService.CreateTournamentStaff(tournamentStaff);
+
+            return Ok();
         }
 
-        [HubRolePermission(HubRole = PlayerHubRole.Administrator)]
         [HttpPut("{id}")]
-        public void UpdateTournamentStaffMember(TournamentStaff tournamentStaff)
+        public IActionResult UpdateTournamentStaffMember(TournamentStaff tournamentStaff)
         {
-            // TODO: MUST BE ADMIN OF TOURNAMENT
+            if (!_tournamentService.IsTournamentOrganiser((int)tournamentStaff.TournamentId, User.GetSteamId()) && !_playerService.IsOwner(User.GetSteamId()))
+            {
+                return Unauthorized();
+            }
+
             _tournamentService.UpdateTournamentStaff(tournamentStaff);
+
+            return Ok();
         }
 
     }
